+++ Rebus/Bus/Advanced/IRoutingApi.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Rebus.Routing;
@@ -18,5 +19,11 @@ namespace Rebus.Bus.Advanced
         /// Sends the message as a routing slip that will visit the destinations specified by the given <see cref="Itinerary"/>
         /// </summary>
         Task SendRoutingSlip(Itinerary itinerary, object message, Dictionary<string, string> optionalHeaders = null);
+
+        /// <summary>
+        /// Explicitly routes the <paramref name="explicitlyRoutedMessage"/> to the destination specified by <paramref name="destinationAddress"/>,
+        /// delaying delivery approximately by the time specified by <paramref name="delay"/>.
+        /// </summary>
+        Task Defer(string destinationAddress, TimeSpan delay, object explicitlyRoutedMessage, Dictionary<string, string> optionalHeaders = null);
     }
 }